Adjust typings.
